+++ core/src/main/java/io/snabble/sdk/ViolationNotification.kt
@@ -1,5 +1,7 @@
 package io.snabble.sdk
 
+import io.snabble.sdk.checkout.ViolationType
+
 /**
  * A notification that a violation occurred
  */
@@ -9,7 +11,7 @@ data class ViolationNotification(
     /** The local generated uuid of the affected item **/
     val refersTo: String?,
     /** The type of the violation **/
-    val type: String? = null,
+    val type: ViolationType? = null,
     /** Non-localized message of the backend which sould be displayed
      * if no localized message is available
      */