+++ core/src/main/java/io/snabble/sdk/checkout/CheckoutState.kt
@@ -110,6 +110,11 @@ enum class CheckoutState {
     INVALID_PRODUCTS,
 
     /**
+     * Invalid items detected. For example if an item can't be found
+     */
+    INVALID_ITEMS,
+
+    /**
      * No payment method available
      */
     NO_PAYMENT_METHOD_AVAILABLE,
@@ -129,4 +134,9 @@ enum class CheckoutState {
      * will not be communicated
      */
     PAYMENT_TRANSFERRED,
+
+    /**
+     * One of the deposit return vouchers added couldn't be redeemed
+     */
+    DEPOSIT_RETURN_REDEMPTION_FAILED
 }