+++ ui/src/main/java/io/snabble/sdk/ui/checkout/Violation.kt
@@ -5,6 +5,7 @@ package io.snabble.sdk.ui.checkout
 import android.content.Context
 import androidx.appcompat.app.AlertDialog
 import io.snabble.sdk.ViolationNotification
+import io.snabble.sdk.checkout.ViolationType
 import io.snabble.sdk.shoppingcart.ShoppingCart
 import io.snabble.sdk.ui.R
 import io.snabble.sdk.ui.utils.I18nUtils
@@ -15,24 +16,20 @@ import io.snabble.sdk.ui.utils.I18nUtils
 fun List<ViolationNotification>.getMessage(context: Context) = joinToString("\n") {
     val res = context.resources
     when (it.type) {
-        "coupon_invalid" -> res.getString(
-            I18nUtils.getIdentifier(res, R.string.Snabble_Violations_couponInvalid),
-            it.name
-        )
+        ViolationType.DEPOSIT_RETURN_ALREADY_REDEEMED ->
+            res.getString(R.string.Snabble_Violations_DepositReturnVoucher_alreadyRedeemed)
 
-        "coupon_currently_not_valid" -> res.getString(
-            I18nUtils.getIdentifier(
-                res,
-                R.string.Snabble_Violations_couponCurrentlyNotValid
-            ), it.name
-        )
+        ViolationType.DEPOSIT_RETURN_DUPLICATED ->
+            res.getString(R.string.Snabble_Violations_DepositReturnVoucher_duplicated)
 
-        "coupon_already_voided" -> res.getString(
-            I18nUtils.getIdentifier(
-                res,
-                R.string.Snabble_Violations_couponAlreadyVoided
-            ), it.name
-        )
+        ViolationType.COUPON_ALREADY_VOIDED ->
+            res.getString(I18nUtils.getIdentifier(res, R.string.Snabble_Violations_couponAlreadyVoided), it.name)
+
+        ViolationType.COUPON_CURRENTLY_NOT_VALID ->
+            res.getString(I18nUtils.getIdentifier(res, R.string.Snabble_Violations_couponCurrentlyNotValid), it.name)
+
+        ViolationType.COUPON_INVALID ->
+            res.getString(I18nUtils.getIdentifier(res, R.string.Snabble_Violations_couponInvalid), it.name)
 
         else -> it.fallbackMessage.orEmpty()
     }